Committing a cache simulator.
I've had it laying around for long time and just got some nosetests to pass.
The simulator is very basic but still usable to make a simple check on
hit/miss ratio when increasing/decreasing the memory of a cache.

There are obviously improvements that can be made such as:
 * Setting a limit on maximum items in the cache.
 * Extending the life of a CacheKeyValue when it is updated with a new value.
 * Speed improvements. I haven't made any profiling at all of this code.
 * More testing.